refactor: 게시물 수정, 삭제로직 수정(#10)
두 로직에 토큰 정보와 일치하지 않으면 오류메세지를 출력하게 수정하였습니다.

+++ src/main/java/gladiator/philosopher/post/service/PostServiceImpl.java
@@ -89,6 +89,9 @@ public class PostServiceImpl implements PostService {
     Post post = postRepository.findById(postId).orElseThrow(
         () -> new CustomException(ExceptionStatus.POST_IS_NOT_EXIST)
     );
+    if (!memberDetails.getMember().getEmail().equals(post.getAccount().getEmail())) {
+      throw new CustomException(ExceptionStatus.UNMATCHED_USER);
+    }
     post.modifyPost(postRequestDto);
     postRepository.save(post);
     return new PostResponseDto(post);
@@ -100,6 +103,9 @@ public class PostServiceImpl implements PostService {
     Post post = postRepository.findById(postId).orElseThrow(
         () -> new CustomException(ExceptionStatus.POST_IS_NOT_EXIST)
     );
+    if (!memberDetails.getMember().getEmail().equals(post.getAccount().getEmail())) {
+      throw new CustomException(ExceptionStatus.UNMATCHED_USER);
+    }
     postRepository.delete(post);
   }
